[#19] refactor : 컴포넌트 분할
인증시간 타이머 변수형태 변경, sendEmail로직 변경
